update heading to its own component
