--- src/components/Easter/SchoolCard.jsx
@@ -1,7 +1,0 @@
-export default function SchoolCard() {
-  return (
-    <>
-      
-    </>
-  )
-}
